Introduce the ParamUtils class
